requirements.txt
asgiref==3.7.1
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2.1
django-multiselectfield==0.1.12
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.1
zope.interface==6.0
